+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### Fixed
 - Fix Åland, Midsummer Day and All Saints Day
+- Fix Canada, FamilyDay
 - Fix Docker Container crash
 - Fix Australia, Add Weekend shift
 ## [1.28.0] - 2021-01-20